requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
django-axes>=6.1  # Brute force protection & rate limiting
bleach>=6.1  # HTML sanitization for user inputs
Pillow>=10.0
qrcode[pil]>=7.4
celery>=5.3
django-celery-beat>=2.6
redis>=5.0
django-redis>=5.4
python-decouple>=3.8
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.6
python-json-logger>=2.0
requests>=2.31.0
django-storages[s3]>=1.14
boto3>=1.34
